requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pillow==10.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
selenium==4.15.2

